Renumber class indexes for determinism
When giving names to classes we sort the classes first.  The goal
being that when we have two classes with the same inferred name
`Foo` we'll end up with one being `Foo` and the other `OtherFoo`
in a deterministic way, and not depending on which one came first
in your schema file, for example.

The problem is that `Eq` and `Ord` on classes is defined such
that it takes class indexes into account when comparing, which
totally depends on which one came first in the schema.

We also can’t trivially do a deep comparison (i.e. follow the
class indexes) because we might at some point want to support
recursive types.

What this does is regenerate the graph starting from the
toplevels in a deterministic order (depth first), so that no
matter what the order we’ll end up with the same indexes.


Original commit bf92279e281bbc68a788a2288ae6b6f5db80756e