Don't re-sort items in transformNames
The whole point of normalizing the order of classes was to not have to do sorting.  In this case, the difference between given/inferred names when coming from JSON vs from Schema tripped us up.   Those names are subject to comparison, too, and from Schema we always get given names, but from JSON they're mostly inferred, so the comparisons work out differently.