Class vs struct option for Swift
The default is back to struct.  We have to run the
JSON Schema tests with class because they produce
recursive types which are not possible with structs.

We should automatically detect cycles in the IR and
generate classes for those, and structs for the rest,
at least by default.  Having this analysis might be
useful for other languages, too.  One other use case
would be forward declarations, eg for C++.