Remove name of Name
The idea of `name` was to identify Names so that
we can allow the user to override their assigned
names.  For that to work they have to be unique, which
they aren't yet, and we don't have this functionality
yet anyway, so there's no point having them around.

Also, we have to hash properly, but we can't add a
Name to its namespace in the Name constructor,
because that will call `hashCode` on the Name, which
isn't properly constructed yet.  Therefore, adding
a Name to a Namespace must happen outside the
constructor.  Also also, the Name doesn't even need
a reference to its Namespace.