AssociatedName
An AssociatedName is associated with a Name, and
is assigned at the same time.  If either of them
conflicts, both are assigned differently again.

If you wanted to have, say, a type preferably called
`Foo`, and an unmarshal method that was called the
same but with the prefix `Unmarshal`, with a
`DependencyType` you might end up with with
`Foo` as the type, but `PurpleUnmarshalFoo` for
the method, because `UnmarshalFoo` was already
taken.  With the `AssociatedType`, since they're
both named at the same time, you'd get `PurpleFoo`
and `UnmarshalPurpleFoo`.