Support alternative names
Names can now have any number of
alternatives, which means we can for example
try naming a type by prefixing it with the
type it's referenced from, before resorting to
random prefixes.  So, for example, in

    { "foo": { "bar": { ... } } }

We could try to name the innermost type
`Bar` by default, and `FooBar` as an
alternative.  Only then would we switch to
`PurpleBar` or whatever.