Make JSON Schema output more deterministic
Re-enable the schema-json tests, where we compare
a generated schema to a schema generated from the
generated schema.  We're using C# now instead of Go
because there's a property `-` in one of the tests, and
Go doesn't handle that.  The class with that property is
supposed to be a map, but we don't detect that yet.