Make TypeRef reference TypeGraph, not TypeBuilder
The original reason for referencing TypeBuilder was
that the TypeRef is dereferenced even during graph
building, at which time the TypeGraph doesn't know
about any nodes, so they must be gotten from the
TypeBuilder.  I rewrote it so the TypeGraph references
the TypeBuilder until it's frozen.  This way we can
completely get rid of the TypeBuilder once the graph
is frozen.