Don't allow date/time strings to be enum types
Previously we could have an enum with cases like, say `2017-12-04`.  With languages that don’t support dates that’s was the case.  But if we generate JSON Schema, that becomes a `date`, and the actual value is lost, so when we generate, say, Swift code from that JSON Schema, the enum is gone and we’re left with a string.  That makes tests that compare JSON->X vs JSON->Schema->X fail.