Make rewriting recursive types less brittle
Previously, a forwarding type ref didn't have an index until
it was resolved when the type it stood for was finished building.
If it was part of a recursive type it could end up in a set or map
before it was finished building, and then it would dereferenced,
which would panic.

This change will create an index for a forwarding type ref in
case it was dereferenced, and we try to make whoever builds the
type use that index.  It could happen that it wants to reuse an
existing type, however, in which case we'll still panic.