Use strings for Go enums. Fixes #373
Unfortunately we don't check the validity of
the enums anymore, both when marshalling
and when unmarshalling.  If we kept the
custom marshalling methods we could check
there, but they wouldn't get invoked in all
cases, like the integers before, so we'd have
checking in most, but not all cases, and a hard
job of explaining which case applies in which
situation.