Separate namespace for forbidden words. Fixes #420
With forbidden words as well as named types and functions in the global namespace, there was no way to make property names that could be the same as global names but still rename them if they clash with forbidden names.  Now we keep forbidden names and global names in separate namespaces.