Markov chain for map inference.  Fixes #3
We're using the Markov chain to come up with the likelihood that
a property name is a "real" property or just a map key.  Then we
lower the threshold number for how many properties we require in
an object to make it a map.  In other words, the more random the
property names in an object, the fewer properties are required to
make it into a map, vs the fixed threshold of 20 we're using now.