Thread type attributes through the whole pipeline
Before this, whenever a graph was rewritten, the new graph would
lose all type attributes except for the names, because we special-
cased adding them.  Now we special-case modifying them, but we
always thread all attributes through.  In addition there's a
mechanism for combining attributes which is used when classes are
combined.  Each attribute kind can specify how its values are to
be combined.