Allow more recursion in JSON Schema
The usage of unification in our JSON Schema input code meant that a type
had to be constructed before it could be incorporated into a union, but
sometimes the union was part of the type, so to construct the type we
would have to construct the union first, but to construct the union we
had to construct the type first.

The fix to this is not to use unification but instead allow more generic
unions, in particular unions containing only one type, as well as unions
containing more than one class and/or map.  We then do a flattening
pass where we reduce everything to the kinds of unions our renderers
support.