Tell GraphRewriteBuilder about unions.  Fixes #526
In the Vega Lite schema we had a case where during graph rewrite
classes were combined which led to one of their properties being
unified into a class, and one of the properties of that class was the
type of the unified class itself, which led to an infinite recursion.

Now when we unify types we let the GraphRewriteBuilder know about
that, and give it a forwarding ref, so that when that is done again
recursively, it resolved to the forwarding ref.