Remove whitespace at end of blank doc comment lines.
Since `commentLineStart` had a space at the end of it, if the
line in the comment was blank, there would be a space at the end
of that blank line.

In Rust, this resulted in a bunch of lines modified by `rustfmt`,
so it would be nicer to just not generate that extra bit of
whitespace.