Get rid of topLevelRefs, do that from the Schema URIs
When before you had to specify

    --add-schema-top-level /definitions/ ./my-schema.json

you can now just specify

    "./my-schema.json#/definitions/"

The rule is that if there's a slash at the end, all the properties
of that object there are taken as types.  If the slash is
missing, the object itself if the type.  The exception is
the root path, which is always taken directly as a type,
whether or not a slash is there, i.e. these are all the same,
and specify just one type:

    ./my-schema.json
    ./my-schema.json#
    ./my-schema.json#/