Repeatedly infer maps until nothing changes
When an inferred map is nested in another inferred map,
the inner one will usually not actually be made into a map
because its original class will be unified with another class
in the process of making the outer map.  On the other
hand, this unification can lead to more inner classes
becoming viable as maps.