Improve Error Messaging for Enumeration Conversion (#2431)
Enhance clarity and informativeness of error messages in the C++ code
generator for enum type conversions. Update the default case in
`to_json` function to include the specific enumeration name and the
unexpected value in the error message. This aids in quicker debugging
and aligns with best practices in error messaging.

Co-authored-by: David Siegel <dvdsgl@users.noreply.github.com>